[AI-694] Extract TranscriptFileClassification helper
Pure refactor. Moves the HEAD-probe classification (New / Partial /
AlreadyLoaded / TooShort / Excluded / ProbeError) out of ImportCommand
so the upcoming ClaudeImportSource and CodexImportSource implementations
can share it. No behavioural change.
