+++ src/Kapacitor.Cli/Commands/ImportCommand.cs
@@ -1,5 +1,4 @@
 using System.Collections.Concurrent;
-using System.Net;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
@@ -340,7 +339,7 @@ static class ImportCommand {
                 .StartAsync(async ctx => {
                         var bar = ctx.AddTask("[yellow]Probing[/]", maxValue: transcriptFiles.Count);
 
-                        var results = await ClassifyAsync(
+                        var results = await TranscriptFileClassification.ClassifyAsync(
                             httpClient,
                             baseUrl,
                             transcriptFiles,
@@ -357,7 +356,7 @@ static class ImportCommand {
             classifications = tmp;
         } else {
             display.Line($"Probing {transcriptFiles.Count} sessions...");
-            classifications = await ClassifyAsync(httpClient, baseUrl, transcriptFiles, minLines, excludedRepos, CancellationToken.None, vendor, excludedPaths: excludedPaths);
+            classifications = await TranscriptFileClassification.ClassifyAsync(httpClient, baseUrl, transcriptFiles, minLines, excludedRepos, CancellationToken.None, vendor, excludedPaths: excludedPaths);
         }
 
         // --- --since filter (Claude path only — Codex pruned at discovery) ---
@@ -1502,252 +1501,6 @@ static class ImportCommand {
         if (c.ExcludedPathKey is { } pathKey && !includedPathKeys.Contains(pathKey)) return true;
 
         return false;
-    }
-
-    internal static async Task<List<SessionClassification>> ClassifyAsync(
-            HttpClient                                                   httpClient,
-            string                                                       baseUrl,
-            List<(string SessionId, string FilePath, string EncodedCwd)> transcripts,
-            int                                                          minLines,
-            string[]?                                                    excludedRepos,
-            CancellationToken                                            ct,
-            string                                                       vendor        = "claude",
-            Action?                                                      onProbed      = null,
-            string[]?                                                    excludedPaths = null
-        ) {
-        using var probeGate = new SemaphoreSlim(8);
-        var       tasks     = new List<Task<SessionClassification>>(transcripts.Count);
-
-        foreach (var (sessionId, filePath, encodedCwd) in transcripts) {
-            tasks.Add(ClassifyOneAsync(httpClient, baseUrl, sessionId, filePath, encodedCwd, minLines, excludedRepos, excludedPaths, probeGate, vendor, onProbed, ct));
-        }
-
-        var results = await Task.WhenAll(tasks);
-
-        return [.. results];
-    }
-
-    static async Task<SessionClassification> ClassifyOneAsync(
-            HttpClient        httpClient,
-            string            baseUrl,
-            string            sessionId,
-            string            filePath,
-            string            encodedCwd,
-            int               minLines,
-            string[]?         excludedRepos,
-            string[]?         excludedPaths,
-            SemaphoreSlim     probeGate,
-            string            vendor,
-            Action?           onProbed,
-            CancellationToken ct
-        ) {
-        try {
-            return await ClassifyOneCoreAsync(httpClient, baseUrl, sessionId, filePath, encodedCwd, minLines, excludedRepos, excludedPaths, probeGate, vendor, ct);
-        } finally {
-            onProbed?.Invoke();
-        }
-    }
-
-    static async Task<SessionClassification> ClassifyOneCoreAsync(
-            HttpClient        httpClient,
-            string            baseUrl,
-            string            sessionId,
-            string            filePath,
-            string            encodedCwd,
-            int               minLines,
-            string[]?         excludedRepos,
-            string[]?         excludedPaths,
-            SemaphoreSlim     probeGate,
-            string            vendor,
-            CancellationToken ct
-        ) {
-        var isCodex = vendor == "codex";
-        var meta    = isCodex ? ExtractCodexSessionMetadata(filePath) : ExtractSessionMetadata(filePath);
-
-        // Short-circuit: kapacitor's own sub-sessions (title / what's-done) never get imported.
-        // Codex rollouts have no analog, so the check is Claude-only.
-        if (!isCodex && TitleGenerator.IsKapacitorSubSession(filePath)) {
-            return new() {
-                SessionId  = sessionId,
-                FilePath   = filePath,
-                EncodedCwd = encodedCwd,
-                Meta       = meta,
-                Status     = ClassificationStatus.InternalSubSession,
-                Vendor     = vendor,
-            };
-        }
-
-        // Codex rollouts carry the session id in two places — the trailing UUID in the
-        // filename (which the rest of the pipeline trusts as the canonical id used for
-        // probe URLs and hook payloads) and `session_meta.payload.id`. Validate they
-        // agree so a renamed/copied file can't import under the wrong server session.
-        if (isCodex && meta.SessionId is { } innerId
-         && Guid.TryParse(innerId, out var innerGuid)
-         && innerGuid.ToString("N") != sessionId) {
-            return new() {
-                SessionId        = sessionId,
-                FilePath         = filePath,
-                EncodedCwd       = encodedCwd,
-                Meta             = meta,
-                Status           = ClassificationStatus.ProbeError,
-                Vendor           = vendor,
-                ProbeErrorReason = "codex session id mismatch (filename vs session_meta.payload.id)",
-            };
-        }
-
-        // Probe the server BEFORE scanning the file. On re-runs the probe returns
-        // 204 (AlreadyLoaded) quickly and we never need to read the transcript.
-        ClassificationStatus status;
-        var                  resumeFromLine   = 0;
-        string?              probeErrorReason = null;
-
-        await probeGate.WaitAsync(ct);
-
-        try {
-            using var resp = await httpClient.GetWithRetryAsync($"{baseUrl}/api/sessions/{sessionId}/last-line", ct: ct);
-
-            switch (resp.StatusCode) {
-                case HttpStatusCode.NotFound:
-                    status = ClassificationStatus.New;
-
-                    break;
-                case HttpStatusCode.NoContent:
-                    status = ClassificationStatus.AlreadyLoaded;
-
-                    break;
-                default:
-                    if (resp.IsSuccessStatusCode) {
-                        var       json = await resp.Content.ReadAsStringAsync(ct);
-                        using var doc  = JsonDocument.Parse(json);
-
-                        if (doc.RootElement.Num("last_line_number") is { } lastLine) {
-                            resumeFromLine = (int)lastLine + 1;
-                            status         = ClassificationStatus.Partial;
-                        } else {
-                            status = ClassificationStatus.AlreadyLoaded;
-                        }
-                    } else {
-                        status           = ClassificationStatus.ProbeError;
-                        probeErrorReason = $"HTTP {(int)resp.StatusCode}";
-                    }
-
-                    break;
-            }
-        } catch (HttpRequestException ex) {
-            status           = ClassificationStatus.ProbeError;
-            probeErrorReason = ex.Message;
-        } finally {
-            probeGate.Release();
-        }
-
-        // Read enough of the local transcript to satisfy two checks at once:
-        //   1. TooShort — fewer lines than minLines.
-        //   2. False Partial — server says last_line_number = N but the local
-        //      transcript has no lines past index N (resumeFromLine would be
-        //      N+1 with nothing to send).
-        // CountLinesUpTo early-exits at the threshold, so the read cost is
-        // bounded by Math.Max(minLines, resumeFromLine + 1) lines.
-        if (status is ClassificationStatus.New or ClassificationStatus.Partial) {
-            var threshold = Math.Max(
-                minLines,
-                status == ClassificationStatus.Partial ? resumeFromLine + 1 : 0
-            );
-
-            if (threshold > 0) {
-                var observedLines = CountLinesUpTo(filePath, threshold);
-
-                if (minLines > 0 && observedLines < minLines) {
-                    return new() {
-                        SessionId  = sessionId,
-                        FilePath   = filePath,
-                        EncodedCwd = encodedCwd,
-                        Meta       = meta,
-                        Status     = ClassificationStatus.TooShort,
-                        TotalLines = observedLines,
-                        Vendor     = vendor,
-                    };
-                }
-
-                // Server has lines >= the local transcript — nothing to resume.
-                if (status == ClassificationStatus.Partial && observedLines <= resumeFromLine) {
-                    status         = ClassificationStatus.AlreadyLoaded;
-                    resumeFromLine = 0;
-                }
-            }
-        }
-
-        // Flag excluded repos/paths for New/Partial sessions. Resolution (include or skip?)
-        // happens later in HandleImport, where we can batch prompts by key.
-        string? excludedRepoKey = null;
-        string? excludedPathKey = null;
-
-        if (status is ClassificationStatus.New or ClassificationStatus.Partial) {
-            var cwd = meta.Cwd ?? SessionImporter.DecodeCwdFromDirName(encodedCwd);
-
-            if (cwd is not null) {
-                if (excludedRepos is { Length: > 0 }) {
-                    var repo = await RepositoryDetection.DetectRepositoryAsync(cwd);
-
-                    if (repo?.Owner is not null && repo.RepoName is not null) {
-                        var key = $"{repo.Owner}/{repo.RepoName}";
-
-                        if (excludedRepos.Contains(key, StringComparer.OrdinalIgnoreCase)) {
-                            excludedRepoKey = key;
-                        }
-                    }
-                }
-
-                if (excludedPathKey is null && excludedPaths is { Length: > 0 }) {
-                    foreach (var entry in excludedPaths) {
-                        if (PathExclusion.IsExcluded(cwd, [entry])) {
-                            excludedPathKey = PathExclusion.Normalize(entry);
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        // TotalLines is only meaningful for TooShort sessions (where we know the exact
-        // count because it's below the threshold). Leave it at 0 for other statuses —
-        // we only read enough of the file to confirm the TooShort filter didn't apply.
-        return new SessionClassification {
-            SessionId        = sessionId,
-            FilePath         = filePath,
-            EncodedCwd       = encodedCwd,
-            Meta             = meta,
-            Status           = status,
-            ResumeFromLine   = resumeFromLine,
-            ProbeErrorReason = probeErrorReason,
-            ExcludedRepoKey  = excludedRepoKey,
-            ExcludedPathKey  = excludedPathKey,
-            Vendor           = vendor,
-        };
-    }
-
-    /// <summary>
-    /// Count transcript lines with an early exit once <paramref name="threshold"/> lines
-    /// have been observed. The caller only needs to distinguish "below threshold" from
-    /// "at or above"; scanning further would be wasted I/O on large transcripts.
-    /// Returns the exact count when below threshold, or exactly <paramref name="threshold"/>
-    /// once the threshold is reached.
-    /// </summary>
-    static int CountLinesUpTo(string path, int threshold) {
-        try {
-            if (!File.Exists(path)) return 0;
-
-            using var stream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
-            using var reader = new StreamReader(stream);
-            var       count  = 0;
-            while (count < threshold && reader.ReadLine() is not null) count++;
-
-            return count;
-        } catch {
-            // On transient I/O errors (locked file, permissions hiccup) treat the
-            // transcript as "not too short" so the caller proceeds to probe/import
-            // rather than silently classifying it as TooShort and skipping forever.
-            return threshold;
-        }
     }
 
     /// <summary>
