+++ test/Kapacitor.Cli.Tests.Unit/Import/ImportClassifyTests.cs
@@ -44,7 +44,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -74,7 +74,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -105,7 +105,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -134,7 +134,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -160,7 +160,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -193,7 +193,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -234,7 +234,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             transcripts,
@@ -261,7 +261,7 @@ public class ImportClassifyTests : IDisposable {
         var probedCount = 0;
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client,
             _server.Url!,
             paths,
@@ -297,7 +297,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client, _server.Url!, transcripts,
             minLines: 15, excludedRepos: null, CancellationToken.None
         );
@@ -326,7 +326,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client, _server.Url!, transcripts,
             minLines: 15, excludedRepos: null, CancellationToken.None
         );
@@ -381,7 +381,7 @@ public class ImportClassifyTests : IDisposable {
 
         using var client = new HttpClient();
 
-        var result = await ImportCommand.ClassifyAsync(
+        var result = await TranscriptFileClassification.ClassifyAsync(
             client, _server.Url!, transcripts,
             minLines: 15,
             excludedRepos: ["any/repo"],
